OS-21 add hibernate config and migration provider

+++ owl-impl/src/main/java/com/geo/owl/building/model/ElevationMarkHistory.java
@@ -29,5 +29,5 @@ public class ElevationMarkHistory {
     private String elevationMarkId;
 
     @Version
-    private int lock;
+    private long lock;
 }
